+++ stdnet/contrib/searchengine/ignore.py
@@ -13,10 +13,11 @@ why,will,with,would,yet,you,your
 
 
 ALPHABET = 'abcdefghijklmnopqrstuvwxyz'
+NUMBERS = '0123456789'
+ALPHA_NUMERIC = ALPHABET+NUMBERS
 
-
-# Do not index any words shorter than this
 MIN_WORD_LENGTH = 3
 
-# Consider these characters to be punctuation (they will be replaced with spaces prior to word extraction)
-PUNCTUATION_CHARS = ".,;:!?@£$%^&*()-–<>[]{}\\|/`~'\""
+# Consider these characters to be punctuation
+# they will be replaced with spaces prior to word extraction
+PUNCTUATION_CHARS = ",.;:'@~#[]{}`!$%^&*()"