+++ stdnet/contrib/searchengine/tests/regression/testsearch/models.py
@@ -1,0 +1,6 @@
+from stdnet import orm
+
+class Item(orm.StdModel):
+    name = orm.SymbolField()
+    content = orm.CharField()
+    counter = orm.IntegerField()